Patch management improvements
Add phing property "check_patches" for Jenkins to call.
Failing patches or already applied patches will fail builds if check_patches is defined.
The ApplyPatchesTask Task now fails if a patch fails and identifies if a patch is already applied.
Bump version to 2.x as this is a breaking change.